requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
jinja2==3.1.4
python-telegram-bot>=21.3
anthropic>=0.40.0
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv>=1.0.0
Pillow>=10.0.0
